// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  org_admin: 'org_admin',
  member: 'member'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Status = {
  todo: 'todo',
  in_progress: 'in_progress',
  review: 'review',
  done: 'done'
} as const

export type Status = (typeof Status)[keyof typeof Status]


export const Priority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent'
} as const

export type Priority = (typeof Priority)[keyof typeof Priority]